+++ andes/routines/tds.py
@@ -49,10 +49,14 @@ class TDS(BaseRoutine):
                                      ('qrt', 0),
                                      ('kqrt', 1.0),
                                      ('store_z', 0),
-                                     ('store_f', 0.0),
-                                     ('store_h', 0.0),
-                                     ('store_i', 0.0),
-                                     ('no_tqdm', 0.0),
+                                     ('store_f', 0),
+                                     ('store_h', 0),
+                                     ('store_i', 0),
+                                     ('limit_store', 0),
+                                     ('max_store', 900),
+                                     ('save_every', 1),
+                                     ('save_mode', 'auto'),
+                                     ('no_tqdm', 0),
                                      )))
         self.config.add_extra("_help",
                               method='DAE solution method',
@@ -77,6 +81,10 @@ class TDS(BaseRoutine):
                               store_f='store RHS of diff. equations',
                               store_h='store RHS of external diff. equations',
                               store_i='store RHS of external algeb. equations',
+                              limit_store='limit in-memory timeseries storage',
+                              max_store='maximum steps of data stored in memory before offloading',
+                              save_every='save results for one step every "save_every" steps',
+                              save_mode='automatically or manually save output data when done',
                               no_tqdm='disable tqdm progressbar and outputs',
                               )
         self.config.add_extra("_alt",
@@ -101,6 +109,10 @@ class TDS(BaseRoutine):
                               store_f=(0, 1),
                               store_h=(0, 1),
                               store_i=(0, 1),
+                              limit_store=(0, 1),
+                              max_store='positive integer',
+                              save_every='integer',
+                              save_mode=('auto', 'manual'),
                               no_tqdm=(0, 1),
                               )
 
@@ -186,6 +198,7 @@ class TDS(BaseRoutine):
         #   to the new array after extending `dae.y`.
         system.set_address(models=system.exist.pflow_tds)
         system.set_dae_names(models=system.exist.tds)
+        system.set_output_subidx(models=system.exist.pflow_tds)
 
         system.dae.clear_ts()
         system.store_sparse_pattern(models=system.exist.pflow_tds)
@@ -344,6 +357,10 @@ class TDS(BaseRoutine):
         self.qrt_start = time.time()
         self.headroom = 0.0
 
+        # write variable list file at the beginning
+        if not system.files.no_output:
+            system.dae.write_lst(self.system.files.lst)
+
         t0, _ = elapsed()
 
         while (system.dae.t - self.h < self.config.tf) and (not self.busted):
@@ -362,17 +379,36 @@ class TDS(BaseRoutine):
                 self.call_stats.append((system.dae.t.tolist(), self.niter, step_status))
 
             if step_status:
-                dae.store()
+                if config.save_every != 0:
+                    if config.save_every == 1:
+                        dae.store()
+                    else:
+                        if dae.kcount % config.save_every == 0:
+                            dae.store()
+
+                # offload if exceeds `max_store`
+                if self.config.limit_store and len(dae.ts._ys) >= self.config.max_store:
+
+                    # write to file if enabled
+                    if not system.files.no_output:
+                        self.save_output()
+                        logger.info("Offload data from memory to file for t=%.2f - %.2f sec",
+                                    dae.ts.t[0], dae.ts.t[-1])
+
+                    # clear storage in memory anyway
+                    dae.ts.reset()
 
                 self.streaming_step()
+
                 if self.check_criteria() is False:
-                    self.err_msg = 'Violated stability criteria.'
+                    self.err_msg = 'Violated stability criteria. To turn off, set [TDS].criteria = 0.'
                     self.busted = True
 
                 # check if the next step is critical time
                 self.do_switch()
                 self.calc_h()
                 dae.t += self.h
+                dae.kcount += 1
 
                 # show progress in percentage
                 perc = max(min((dae.t - config.t0) / (config.tf - config.t0) * 100, 100), 0)
@@ -420,15 +456,24 @@ class TDS(BaseRoutine):
 
         t1, s1 = elapsed(t0)
         self.exec_time = t1 - t0
-        logger.info('Simulation completed in %s.', s1)
+        logger.info('Simulation to t=%.2f sec completed in %s.', config.tf, s1)
 
         if config.qrt:
             logger.debug('QRT headroom time: %.4g s.', self.headroom)
 
-        # need to unpack data in case of resumed simulations.
-        system.dae.ts.unpack()
-        if not system.files.no_output:
+        # in case of resumed simulations,
+        # manually unpack data to update arrays in `dae.ts`
+        # disable warning in case data has just been dumped
+        system.dae.ts.unpack(warn_empty=False)
+
+        if (not system.files.no_output) and (config.save_mode == 'auto'):
+            t0, _ = elapsed()
             self.save_output()
+            _, s1 = elapsed(t0)
+
+            np_file = self.system.files.npz
+            logger.info('Outputs to "%s" and "%s".', self.system.files.lst, np_file)
+            logger.info('Outputs written in %s.', s1)
 
         # end data streaming
         if system.config.dime_enabled:
@@ -668,21 +713,13 @@ class TDS(BaseRoutine):
             True if files are written. False otherwise.
         """
 
-        t0, _ = elapsed()
-
-        self.system.dae.write_lst(self.system.files.lst)
-
         if npz is True:
-            np_file = self.system.files.npz
             self.system.dae.write_npz(self.system.files.npz)
         else:
-            np_file = self.system.files.npy
             self.system.dae.write_npy(self.system.files.npy)
 
-        _, s1 = elapsed(t0)
+        self.system.dae.ts.idx_ptr = len(self.system.dae.ts.t)
 
-        logger.info('Outputs to "%s" and "%s".', self.system.files.lst, np_file)
-        logger.info('Outputs written in %s.', s1)
         return True
 
     def do_switch(self):